+++ README.md
@@ -9,10 +9,10 @@ The suggestions should be restricted to cities in the USA and Canada with a popu
 - the partial (or complete) search term is passed as a query string parameter `q`
 - the caller's location can optionally be supplied via query string parameters `latitude` and `longitude` to help improve relative scores
 - the endpoint returns a JSON response with an array of scored suggested matches
-    - the suggestions are sorted by descending score
-    - each suggestion has a score between 0 and 1 (inclusive) indicating confidence in the suggestion (1 is most confident)
-    - each suggestion has a name which can be used to disambiguate between similarly named locations
-    - each suggestion has a latitude and longitude
+  - the suggestions are sorted by descending score
+  - each suggestion has a score between 0 and 1 (inclusive) indicating confidence in the suggestion (1 is most confident)
+  - each suggestion has a name which can be used to disambiguate between similarly named locations
+  - each suggestion has a latitude and longitude
 - all functional tests should pass (additional tests may be implemented as necessary).
 - the final application should be [deployed to Heroku](https://devcenter.heroku.com/articles/getting-started-with-nodejs).
 - feel free to add more features if you like!
@@ -66,7 +66,6 @@ These responses are meant to provide guidance. The exact values can vary based o
 }
 ```
 
-
 ### Non-functional
 
 - All code should be written in Javascript, Typescript or PHP.
@@ -108,7 +107,7 @@ You are going to need:
 ### Setting up your environment
 
 1. Begin by forking this repo and cloning your fork. GitHub has apps for [Mac](http://mac.github.com/) and
-[Windows](http://windows.github.com/) that make this easier.
+   [Windows](http://windows.github.com/) that make this easier.
 
 2. Install [nvm](https://github.com/nvm-sh/nvm#install--update-script) or your preferred node version manager.
