+++ data/datasource.js
@@ -1,6 +1,10 @@
+/**
+ * DatasourceInterface intent is to mock an Interface to support different implementations for getting
+ * data.
+ */
 class DatasourceInterface {
   constructor() {
-    if (!this.getData) throw new Error("datasource must return data");
+    if (!this.getData) throw new Error("getData must be implemented");
   }
 }
