+++ data/file-datasource.js
@@ -2,12 +2,19 @@ import { parse } from "csv";
 import fs from "fs";
 import DatasourceInterface from "./datasource.js";
 
-class FileDatasource extends DatasourceInterface {
+/**
+ * FileDatasource: implementation of DatasourceInterface to extract data from a file path.
+ */
+export default class FileDatasource extends DatasourceInterface {
   constructor(path) {
     super();
     this.path = path;
   }
 
+  /**
+   * getData: parses the data from a tsv file.
+   * @returns Promise<Array>
+   */
   getData() {
     return new Promise((resolve, reject) => {
       const parser = parse(
@@ -24,5 +31,3 @@ class FileDatasource extends DatasourceInterface {
     });
   }
 }
-
-export default FileDatasource;
